1-26-12
We tried finding the solution to the login page.
We now have login page that will take the data from the database and render it into our view files with username and id passed from the controller file.
It also checks the validation of username and password and redirects it accordingly.
We also modified the controller page of site/index to redirect to our login page.

The updated file should be available in our github repo.

Next meeting will be on Tuesday 31st Jan after class.

Thanks,
-Yojan
   - and Alex!